Redesign BaseClientDatabase as an ABCMeta
